conviction and evaluate
