+++ howso/direct/client.py
@@ -3517,69 +3517,6 @@ class HowsoDirectClient(AbstractHowsoClient):
 
         return ret
 
-    def get_feature_conviction(
-        self,
-        trainee_id: str,
-        *,
-        features: Optional[Iterable[str]] = None,
-        action_features: Optional[Iterable[str]] = None,
-        familiarity_conviction_addition: bool = True,
-        familiarity_conviction_removal: bool = False,
-        weight_feature: Optional[str] = None,
-        use_case_weights: bool = False
-    ) -> Dict:
-        """
-        Get familiarity conviction for features in the model.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The id of the trainee.
-        features : iterable of str, optional
-            An iterable of feature names to calculate convictions. At least 2
-            features are required to get familiarity conviction. If not
-            specified all features will be used.
-        action_features : iterable of str, optional
-            An iterable of feature names to be treated as action features
-            during conviction calculation in order to determine the conviction
-            of each feature against the set of action_features. If not
-            specified, conviction is computed for each feature against the rest
-            of the features as a whole.
-        familiarity_conviction_addition : bool, default True
-            Calculate and output familiarity conviction of adding the
-            specified features in the output.
-        familiarity_conviction_removal : bool, default False
-            Calculate and output familiarity conviction of removing
-            the specified features in the output.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-
-        Returns
-        -------
-        dict
-            A dict with familiarity_conviction_addition or
-            familiarity_conviction_removal
-        """
-        self._auto_resolve_trainee(trainee_id)
-        validate_list_shape(features, 1, "features", "str")
-        validate_list_shape(action_features, 1, "action_features", "str")
-        if self.verbose:
-            print('Getting conviction of features for trainee with id: '
-                  f'{trainee_id}')
-        return self.howso.compute_conviction_of_features(
-            trainee_id,
-            features=features,
-            action_features=action_features,
-            familiarity_conviction_addition=familiarity_conviction_addition,
-            familiarity_conviction_removal=familiarity_conviction_removal,
-            weight_feature=weight_feature,
-            use_case_weights=use_case_weights
-        )
-
     def add_feature(
         self,
         trainee_id: str,
@@ -5399,49 +5336,3 @@ class HowsoDirectClient(AbstractHowsoClient):
 
         self.howso.analyze(trainee_id, **analyze_params)
         self._auto_persist_trainee(trainee_id)
-
-    def evaluate(
-        self,
-        trainee_id: str,
-        features_to_code_map: Dict[str, str],
-        *,
-        aggregation_code: Optional[str] = None
-    ) -> Dict[str, Dict[str, object]]:
-        """
-        Evaluate custom code on feature values of all cases in the trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee.
-        features_to_code_map : dict of str to str
-            A dictionary with feature name keys and custom Amalgam code string values.
-
-            The custom code can use "#feature_name 0" to reference the value
-            of that feature for each case.
-        aggregation_code : str, optional
-            A string of custom Amalgam code that can access the list of values
-            derived form the custom code in features_to_code_map.
-            The custom code can use "#feature_name 0" to reference the list of
-            values derived from using the custom code in features_to_code_map.
-
-        Returns
-        -------
-        dict
-            A dictionary with keys: 'evaluated' and 'aggregated'
-
-            'evaluated' is a dictionary with feature name
-            keys and lists of values derived from the features_to_code_map
-            custom code.
-
-            'aggregated' is None if no aggregation_code is given, it otherwise
-            holds the output of the custom 'aggregation_code'
-        """
-        self._auto_resolve_trainee(trainee_id)
-
-        evaluate_params = dict(
-            features_to_code_map=features_to_code_map,
-            aggregation_code=aggregation_code,
-        )
-
-        return self.howso.evaluate(trainee_id, **evaluate_params)
